Add a sub_query query builder to the DataSet class.
