+++ fireant/queries/__init__.py
@@ -1,5 +1,6 @@
 from .builder import (
     DataSetQueryBuilder,
+    DataSetSubQueryBuilder,
     DimensionChoicesQueryBuilder,
     DimensionLatestQueryBuilder,
 )
