+++ fireant/tests/queries/test_build_joins.py
@@ -51,8 +51,7 @@ class QueryBuilderJoinTests(TestCase):
                          '"candidate_id" "$candidate_id",'
                          '"candidate_spend" "$candidate_spend" '
                          'FROM "politics"."politician_spend" '
-                         'GROUP BY "$candidate_id","$candidate_spend" '
-                         'ORDER BY "$candidate_id","$candidate_spend"'
+                         'GROUP BY "$candidate_id","$candidate_spend"'
                          ') "politician_spend" '
                          'ON "politician"."id"="politician_spend"."candidate_id" '
                          'GROUP BY "$timestamp","$candidate-id","$average-candidate-spend-per-candidacy" '
